Enum alloy_rpc_types_engine::payload::PayloadError

pub enum PayloadError {

    ExtraData(Bytes),
    BaseFee(U256),
    BlobGasUsed(U256),
    ExcessBlobGas(U256),
    PreShanghaiBlockWithWitdrawals,
    PostShanghaiBlockWithoutWitdrawals,
    PreCancunBlockWithBlobTransactions,
    PreCancunBlockWithBlobGasUsed,
    PreCancunBlockWithExcessBlobGas,
    PreCancunWithCancunFields,
    PostCancunBlockWithoutBlobTransactions,
    PostCancunBlockWithoutBlobGasUsed,
    PostCancunBlockWithoutExcessBlobGas,
    PostCancunWithoutCancunFields,
    BlockHash {
        execution: B256,
        consensus: B256,
    },
    InvalidVersionedHashes,
    Decode(Error),

}

Error that can occur when handling payloads.

Variants

ExtraData(Bytes)

Invalid payload extra data.

BaseFee(U256)

Invalid payload base fee.

BlobGasUsed(U256)

Invalid payload blob gas used.

ExcessBlobGas(U256)

Invalid payload excess blob gas.

PreShanghaiBlockWithWitdrawals

withdrawals present in pre-shanghai payload.

PostShanghaiBlockWithoutWitdrawals

withdrawals missing in post-shanghai payload.

PreCancunBlockWithBlobTransactions

blob transactions present in pre-cancun payload.

PreCancunBlockWithBlobGasUsed

blob gas used present in pre-cancun payload.

PreCancunBlockWithExcessBlobGas

excess blob gas present in pre-cancun payload.

PreCancunWithCancunFields

cancun fields present in pre-cancun payload.

PostCancunBlockWithoutBlobTransactions

blob transactions missing in post-cancun payload.

PostCancunBlockWithoutBlobGasUsed

blob gas used missing in post-cancun payload.

PostCancunBlockWithoutExcessBlobGas

excess blob gas missing in post-cancun payload.

PostCancunWithoutCancunFields

cancun fields missing in post-cancun payload.

BlockHash

Invalid payload block hash.

Fields

execution: B256

The block hash computed from the payload.

consensus: B256

The block hash provided with the payload.

InvalidVersionedHashes

Expected blob versioned hashes do not match the given transactions.

Decode(Error)

Encountered decoding error.

Implementations

impl PayloadError

pub const fn is_block_hash_mismatch(&self) -> bool

Returns true if the error is caused by a block hash mismatch.

pub const fn is_invalid_versioned_hashes(&self) -> bool

Returns true if the error is caused by invalid block hashes (Cancun).