Struct alloy_rpc_types_engine::payload::BlobsBundleV1

pub struct BlobsBundleV1 {
    pub commitments: Vec<Bytes48>,
    pub proofs: Vec<Bytes48>,
    pub blobs: Vec<Blob>,
}

This includes all bundled blob related data of an executed payload.

Fields
commitments: Vec<Bytes48>

All commitments in the bundle.

proofs: Vec<Bytes48>

All proofs in the bundle.

blobs: Vec<Blob>

All blobs in the bundle.

Implementations

impl BlobsBundleV1

pub fn new(sidecars: impl IntoIterator<Item = BlobTransactionSidecar>) -> Self

Creates a new blob bundle from the given sidecars.

This folds the sidecar fields into single commit, proof, and blob vectors.

pub fn take(&mut self, len: usize) -> (Vec<Bytes48>, Vec<Bytes48>, Vec<Blob>)

Take len blob data from the bundle.

Panics

If len is more than the blobs bundle len.

pub fn pop_sidecar(&mut self, len: usize) -> BlobTransactionSidecar

Returns the sidecar from the bundle

Panics

If len is more than the blobs bundle len.