Crate alloy_rpc_types_engine

alloy-rpc-types-engine

Types for the engine Ethereum JSON-RPC namespace.

Engine API types: https://github.com/ethereum/execution-apis/blob/main/src/engine/authentication.md and https://eips.ethereum.org/EIPS/eip-3675, following the execution specs https://github.com/ethereum/execution-apis/tree/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine.

Re-exports

pub use self::payload::*;

Modules

cancun 🔒

Contains types related to the Cancun hardfork that will be used by RPC to communicate with the beacon consensus engine.

forkchoice 🔒

identification 🔒

Client identification: https://github.com/ethereum/execution-apis/blob/main/src/engine/identification.md

jwt 🔒

JWT (JSON Web Token) utilities for the Engine API.

optimism 🔒

payload

Payload types.

transition 🔒

Structs

CancunPayloadFields

Fields introduced in engine_newPayloadV3 that are not present in the ExecutionPayload RPC object.

Claims

Claims in JWT are used to represent a set of information about an entity. Claims are essentially key-value pairs that are encoded as JSON objects and included in the payload of a JWT. They are used to transmit information such as the identity of the entity, the time the JWT was issued, and the expiration time of the JWT, among others.

ClientVersionV1

Contains information which identifies a client implementation.

DepositRequestV1

This structure maps onto the deposit object from EIP-6110.

ForkchoiceState

This structure encapsulates the fork choice state

ForkchoiceUpdated

Represents a successfully processed forkchoice state update.

JwtSecret

Value-object holding a reference to a hex-encoded 256-bit secret key. A JWT secret key is used to secure JWT-based authentication. The secret key is a shared secret between the server and the client and is used to calculate a digital signature for the JWT, which is included in the JWT along with its payload.

MaybeCancunPayloadFields

A container type for CancunPayloadFields that may or may not be present.

OptimismExecutionPayloadEnvelopeV3

This structure maps for the return value of engine_getPayload of the beacon chain spec, for V3.

OptimismExecutionPayloadEnvelopeV4

This structure maps for the return value of engine_getPayload of the beacon chain spec, for V4.

OptimismPayloadAttributes

Optimism Payload Attributes

TransitionConfiguration

This structure contains configurable settings of the transition process.

WithdrawalRequestV1

Represents an execution layer triggerable withdrawal request.

Enums

ClientCode

This enum defines a standard for specifying a client with just two letters. Clients teams which have a code reserved in this list MUST use this code when identifying themselves.

ForkchoiceUpdateError

A standalone forkchoice update errors for RPC.

JwtError

Errors returned by the JwtSecret

Constants

CAPABILITIES

The list of all supported Engine capabilities available over the engine endpoint.

INVALID_FORK_CHOICE_STATE_ERROR

invalid forkchoice state error code.

INVALID_FORK_CHOICE_STATE_ERROR_MSG

invalid forkchoice state error message.

INVALID_PAYLOAD_ATTRIBUTES_ERROR

invalid payload attributes error code.

INVALID_PAYLOAD_ATTRIBUTES_ERROR_MSG

invalid payload attributes error message.

Type Aliases

ForkChoiceUpdateResult

Represents possible variants of a processed forkchoice update.